# Adds parquet_column_metadata function