--- a/extension/parquet/include/parquet_metadata.hpp
+++ b/extension/parquet/include/parquet_metadata.hpp
@@ -38,4 +38,9 @@ public:
 	ParquetBloomProbeFunction();
 };
 
+class ParquetColumnMetadataFunction : public TableFunction {
+public:
+	ParquetColumnMetadataFunction();
+};
+
 } // namespace duckdb