--- a/extension/parquet/include/parquet_metadata.hpp
+++ b/extension/parquet/include/parquet_metadata.hpp
@@ -38,4 +38,9 @@ public:
 	ParquetBloomProbeFunction();
 };
 
+class ParquetFullMetadataFunction : public TableFunction {
+public:
+	ParquetFullMetadataFunction();
+};
+
 } // namespace duckdb