Following up from https://github.com/duckdb/ducklake/pull/423 adding this function greatly reduces the time to ingest files to ducklake, this cannot be done easily in pure sql because the types of the metadata are file specific and must be cast from string to the correct type before taking the min / max.

The goal is to make this function generally usable, not only targeting the metrics that ducklake uses.

I tried adding some extra columns other than the pure ones ducklake needs, but am not sure which aggregations make the most sense.

All are basically free relative to doing it in explicit C++ within ducklake due the costs of returning  files x columns x row_groups rows vs files x columns rows and the ability to paralellize the aggregations themselves. There  is also a benefit with this function that we do fewer String -> Value conversions than within the Merge logic of ducklake.

For reference ingest times in ducklake:
With this optimization and preserve insertion order false:
```
name	run	timing
benchmark/ingest/add_files_lineitem.benchmark	1	0.353124
benchmark/ingest/add_files_lineitem.benchmark	2	0.325152
benchmark/ingest/add_files_lineitem.benchmark	3	0.336028
benchmark/ingest/add_files_lineitem.benchmark	4	0.334177
benchmark/ingest/add_files_lineitem.benchmark	5	0.331770
```

Without this optimization but with all others:
```
name	run	timing
benchmark/ingest/add_files_lineitem.benchmark	1	1.085267
benchmark/ingest/add_files_lineitem.benchmark	2	1.046103
benchmark/ingest/add_files_lineitem.benchmark	3	1.047027
benchmark/ingest/add_files_lineitem.benchmark	4	1.035097
benchmark/ingest/add_files_lineitem.benchmark	5	1.039195
```

The benchmark understates the performance impact because the dataset generated is high in files and relatively low number of row groups, on files with more row groups per file the impact is much more significant.
On use with actual files (not tiny ones created for benchmark purposes)
- With this optimization and preserve insertion order false: 3.5 seconds
- With this optimization and preserve insertion order true: 13.6 seconds
- Without this optimization: 24 seconds

### Questions:
1. Is adding this function worth considering?
2. If so what column level metadata aggregations might be useful?

Waiting on adding explicit tests until the feasibility is established