@Mytherin 2025-09-25
Thanks for the PR!

I think this is a bit too specific for DuckLake - given these properties can also be extracted from the parquet metadata functions in a different manner.

I do have a different idea in this space. I haven't done that much profiling here - but intuitively the by far dominant cost here should be reading of the metadata, i.e. the I/O and thrift decoding of the Parquet metadata. 

There's room to improve that - in particular, in `ducklake_add_data_files`, we are reading the same Parquet metadata multiple times:

```sql
SELECT file_name, name, type, num_children, converted_type, scale, precision, field_id, logical_type
FROM parquet_schema(%s)

SELECT file_name, column_id, num_values, coalesce(stats_min, stats_min_value), coalesce(stats_max, stats_max_value), stats_null_count, total_compressed_size, geo_bbox, geo_types
FROM parquet_metadata(%s)

SELECT file_name, num_rows, footer_size, file_size_bytes
FROM parquet_file_metadata(%s)
```

What if, instead of having these be separate function calls, we could get all of this metadata in one go?

My idea would be to introduce a new function - `parquet_full_metadata` - that would return all Parquet metadata as a set of nested structures. The structure would be effectively a union of the current metadata functions:

```sql
parquet_file_metadata STRUCT(file_name VARCHAR, created_by VARCHAR, num_rows BIGINT, num_row_groups BIGINT, format_version BIGINT, encryption_algorithm VARCHAR, footer_signing_key_metadata VARCHAR, file_size_bytes UBIGINT, footer_size UBIGINT)

parquet_metadata STRUCT(file_name VARCHAR, row_group_id BIGINT, row_group_num_rows BIGINT, row_group_num_columns BIGINT, row_group_bytes BIGINT, column_id BIGINT, file_offset BIGINT, num_values BIGINT, path_in_schema VARCHAR, "type" VARCHAR, stats_min VARCHAR, stats_max VARCHAR, stats_null_count BIGINT, stats_distinct_count BIGINT, stats_min_value VARCHAR, stats_max_value VARCHAR, "compression" VARCHAR, encodings VARCHAR, index_page_offset BIGINT, dictionary_page_offset BIGINT, data_page_offset BIGINT, total_compressed_size BIGINT, total_uncompressed_size BIGINT, key_value_metadata MAP(BLOB, BLOB), bloom_filter_offset BIGINT, bloom_filter_length BIGINT, min_is_exact BOOLEAN, max_is_exact BOOLEAN, row_group_compressed_bytes BIGINT, geo_bbox STRUCT(xmin DOUBLE, xmax DOUBLE, ymin DOUBLE, ymax DOUBLE, zmin DOUBLE, zmax DOUBLE, mmin DOUBLE, mmax DOUBLE), geo_types VARCHAR[])[]

parquet_schema STRUCT(file_name VARCHAR, "name" VARCHAR, "type" VARCHAR, type_length VARCHAR, repetition_type VARCHAR, num_children BIGINT, converted_type VARCHAR, scale BIGINT, "precision" BIGINT, field_id BIGINT, logical_type VARCHAR, duckdb_type VARCHAR, column_id BIGINT)[]

parquet_kv_metadata STRUCT(file_name VARCHAR, "key" BLOB, "value" BLOB)[]
```

That way we could get all metadata of all files in one go, without having to do multiple passes over the files. The code between `parquet_full_metadata` should lean fully on the existing code for the other functions to ensure we're not duplicating any logic. Another benefit here is that there's no need to join anything again - since we get all of the metadata for one file all grouped together, we can immediately do the computation for that file.