@J-Meyers 2025-10-14
Implemented the new function, I haven't integrated with ducklake yet. So tdb, on those other points.

This had to touch more code than I would have liked to switch from taking in a DataChunk to a more generic form that supports writing to the children of a struct vector.

So what before operated on a `DataChunk` now takes in a `vector<Vector*>` this is compatible with both since `DataChunk::data` is `vector<Vector>` and `StructVector::GetEntries` returns `vector<unique_ptr<Vector>>&`

This forced changing from `output.SetValue(0, ...)` to `output[0]->SetValue(...)`